requirements.txt
anyjson==0.3.3
atomicwrites==1.3.0
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
coverage==4.5.2
Django==2.1.5
django-admin-generator==2.1.0
django-appconf==1.0.2
django-ckeditor==5.6.1
django-environ==0.4.5
django-extensions==2.1.5
django-image-cropping==1.2.0
django-js-asset==1.1.0
django-utils2==2.5.0
easy-thumbnails==2.5
idna==2.8
more-itertools==5.0.0
Pillow==5.4.1
pluggy==0.8.1
polib==1.1.0
psycopg2==2.7.7
psycopg2-binary==2.7.6.1
py==1.7.0
pytest==4.1.1
pytest-cov==2.6.1
pytest-django==3.4.6
python-utils==2.3.0
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
